moved files into /pages
