+++ src/pages/RatingPage.jsx
@@ -1,5 +1,5 @@
 import React, { useState, useEffect } from 'react'
-import api from './utils/api'
+import api from '../utils/api'
 
 function RatingPage() {
   const [searchQuery, setSearchQuery] = useState('')
